net: dsa: sja1105: don't use burst SPI reads for port statistics
The current internal sja1105 driver API is optimized for retrieving many
statistics counters at once. But the switch does not do atomic snapshotting
for them anyway.

There is a movement to start reporting hardware port counters through
ndo_get_stats64 as well, not just ethtool, in order to let the stack
have a better idea of port statistics for offloading interfaces, as well
as for naming standardization. But for ndo_get_stats64, we won't be
reporting based on the hardware ranges, but based on the contents of
struct rtnl_link_stats64, which we'll be populating based on an
aggregate of disjoint hardware counters. So burst reads will not be
helpful there.

Additionally, since Arnd Bergmann's commit ae1804de93f6 ("dsa: sja1105:
dynamically allocate stats structure"), sja1105_get_ethtool_stats
allocates memory dynamically, since struct sja1105_port_status was
deemed to consume too much stack memory. That is far from ideal.
The large structure is only needed because of the burst read. If we read
statistics one by one, we can consume less memory, and we can avoid
dynamic allocation.

Additionally, latency-sensitive interfaces such as PTP operations (for
phc2sys) might suffer if the SPI mutex is being held for too long, which
happens in the case of SPI burst reads. By reading counters one by one,
we give a chance for higher priority processes to preempt and take the
SPI bus mutex for accessing the PTP clock.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>