net: dsa: kill .port_egress_floods overengineering
The bridge offloads the port flags through a single bit mask using
switchdev, which among others, contains learning and flooding settings.

The commit 57652796aa97 ("net: dsa: add support for bridge flags")
missed one crucial aspect of the SWITCHDEV_ATTR_ID_PORT_BRIDGE_FLAGS API
when designing the API one level lower, towards the drivers.
This is that the bitmask of passed brport flags never has more than one
bit set at a time. On the other hand, the prototype passed to the driver
is .port_egress_floods(int port, bool unicast, bool multicast), which
configures two flags at a time.

If the bridge wants to see what flags are supported by the device, it
emits a SWITCHDEV_ATTR_ID_PORT_PRE_BRIDGE_FLAGS attribute. DSA currently
checks if .port_egress_floods is implemented, and if it is, reports both
BR_FLOOD and BR_MCAST_FLOOD. So the driver has no choice if it wants to
inform the bridge that it can configure unicast flooding but not
multicast flooding - the DSA mid layer is standing in the way.

Secondly, as outlined in the first paragraph, flooding is not all that
is passed through the brport flags bitmask, but also learning. If a DSA
switch wanted to configure its learning parameter in spirit of the
current API, new logic would need to be added to DSA for
.port_set_learning. This implies that dsa_port_pre_bridge_flags would
have to deduce, based on which of .port_set_learning and
.port_egress_floods are implemented, what to mark as supported, and then
figure out which of .port_set_learning or .port_egress_floods to call
from dsa_port_bridge_flags. The latter is non-trivial because, as
mentioned, all brport flags finally end up offloaded in the same bitmask,
and ideally we wouldn't want DSA to call an unneeded method, i.e.
.port_set_learning when nothing learning-related has changed, or
.port_egress_floods when nothing flooding-related has changed,
especially when the reconfiguration for that might involve a switch reset.
So to avoid that, DSA's only option would be to perform 'edge detection'
by keeping an unsigned long dsa_port::brport_flags and check for the XOR
between the cached value and the new flags in order to determine what
changed. Either that, or make the drivers check early if anything has
changed, and return if not. But either way, this is fairly overengineered.

Thirdly, currently DSA deems the driver too dumb to deserve knowing that
a SWITCHDEV_ATTR_ID_BRIDGE_MROUTER attribute was offloaded, because it
just calls .port_egress_floods for the CPU port. When we'll add support
for the plain SWITCHDEV_ATTR_ID_PORT_MROUTER, that will become a real
problem because the flood settings will need to be held statefully in
the DSA middle layer, otherwise changing the mrouter port attribute will
impact the flooding attribute. And that's _assuming_ that the underlying
hardware doesn't have anything else to do when a multicast router
attaches to a port than flood unknown traffic to it. If it does, there
will need to be a dedicated .port_set_mrouter anyway.

Lastly, we have DSA drivers that have a backlink into a pure switchdev
driver (felix -> ocelot). It seems reasonable that the other switchdev
drivers should not have to suffer from the oddities of DSA overengineering,
so keeping DSA a pass-through layer makes more sense there.

To simplify the brport flags situation we just delete .port_egress_floods
and we introduce .port_pre_bridge_flags and .port_bridge_flags which are
passed to the driver. Also, the logic from dsa_port_mrouter is removed
and a .port_set_mrouter is created. The .port_pre_bridge_flags callback
is temporary and will be removed in a later patch, together with
SWITCHDEV_ATTR_ID_PORT_PRE_BRIDGE_FLAGS.

Functionally speaking, we simply move the calls to .port_egress_floods
one step lower, in the two drivers that implement it: mv88e6xxx and b53,
so things should work just as before.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>